pyproject.toml
[project]
name = "tfmamba"
version = "0.1.0"
description = "Time–Frequency Dual-Branch State Space Network"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "numpy>=2.3.5",
    "swanlab>=0.7.3",
    "loguru>=0.7.3",
    "scikit-learn>=1.8.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
